# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kledo App",
        "version": "0.1"
    },
    "paths": {
        "/api/employee": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "gf": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "example title"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "example content"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/employee/store": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "content": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Full name",
                                    "salary": 2000000
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Added Employee Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "The name has already been taken."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/settings": {
            "patch": {
                "tags": [
                    "Settings"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "key": {
                                                "type": "string"
                                            },
                                            "value": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "key": "overtime_method",
                                    "value": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Update setting successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "value": {
                                                        "type": "string",
                                                        "example": "TThe selected value is invalid."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/overtimes": {
            "post": {
                "tags": [
                    "Overtimes"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "employee_id": {
                                                "type": "string"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "time_started": {
                                                "type": "string",
                                                "format": "time"
                                            },
                                            "time_ended": {
                                                "type": "string",
                                                "format": "time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "employee_id": "1",
                                    "date": "2022-11",
                                    "time_started": "00:00",
                                    "time_ended": "02:00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Added overtime successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "example": "The date has already been taken."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/overtime-pays/calculate": {
            "get": {
                "tags": [
                    "Overtimes"
                ],
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "YYYY-MM",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Bardi"
                                                    },
                                                    "salary": {
                                                        "type": "integer",
                                                        "example": 2000000
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2022-11-19T02:46:39.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2022-11-19T02:46:39.000000Z"
                                                    },
                                                    "overtime_duration": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "amount": {
                                                        "type": "integer",
                                                        "example": 30000
                                                    },
                                                    "overtimes": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "employee_id": {
                                                                    "type": "string",
                                                                    "example": "1"
                                                                },
                                                                "date": {
                                                                    "type": "string",
                                                                    "example": "09:00:00"
                                                                },
                                                                "time_started": {
                                                                    "type": "string",
                                                                    "example": "09:00:00"
                                                                },
                                                                "time_ended": {
                                                                    "type": "string",
                                                                    "example": "2022-11-19T02:46:39.000000Z"
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "example": "2022-11-19T02:46:44.000000Z"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "example": "2022-11-19T02:46:44.000000Z"
                                                                },
                                                                "overtime_duration": {
                                                                    "type": "integer",
                                                                    "example": 2
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "example": "The date does not match the format Y-m."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}